examples/playground/requirements.txt
langgraph>=1.0
langchain-core>=0.3
langchain-openai>=0.3
websockets>=12.0
python-dotenv>=1.0
